Extract 'GetMeta' as a pure helper to ease metadata migration to FX
