+++ pkg/collector/metadata/agentchecks/agentchecks.go
@@ -9,12 +9,13 @@ import (
 	"context"
 	"encoding/json"
 
+	hostMetadataUtils "github.com/DataDog/datadog-agent/comp/metadata/host/utils"
 	"github.com/DataDog/datadog-agent/pkg/autodiscovery"
 	"github.com/DataDog/datadog-agent/pkg/collector"
 	"github.com/DataDog/datadog-agent/pkg/collector/runner/expvars"
+	"github.com/DataDog/datadog-agent/pkg/config"
 	"github.com/DataDog/datadog-agent/pkg/metadata/common"
 	"github.com/DataDog/datadog-agent/pkg/metadata/externalhost"
-	"github.com/DataDog/datadog-agent/pkg/metadata/host"
 	"github.com/DataDog/datadog-agent/pkg/status"
 	"github.com/DataDog/datadog-agent/pkg/util/hostname"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
@@ -79,7 +80,7 @@ func GetPayload(ctx context.Context) *Payload {
 	}
 
 	// Grab the non agent checks information
-	metaPayload := host.GetMeta(ctx, hostnameData)
+	metaPayload := hostMetadataUtils.GetMetaFromCache(ctx, config.Datadog)
 	metaPayload.Hostname = hostnameData.Hostname
 	cp := common.GetPayload(hostnameData.Hostname)
 	ehp := externalhost.GetPayload()
