+++ pkg/metadata/host/host.go
@@ -21,14 +21,11 @@ import (
 	"github.com/DataDog/datadog-agent/pkg/metadata/host/container"
 	"github.com/DataDog/datadog-agent/pkg/metadata/inventories"
 	"github.com/DataDog/datadog-agent/pkg/otlp"
-	"github.com/DataDog/datadog-agent/pkg/util"
 	"github.com/DataDog/datadog-agent/pkg/util/cache"
 	"github.com/DataDog/datadog-agent/pkg/util/cloudproviders"
-	"github.com/DataDog/datadog-agent/pkg/util/ec2"
 	"github.com/DataDog/datadog-agent/pkg/util/flavor"
 	"github.com/DataDog/datadog-agent/pkg/util/hostname"
 	httputils "github.com/DataDog/datadog-agent/pkg/util/http"
-	"github.com/DataDog/datadog-agent/pkg/util/kubelet"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
 
 	yaml "gopkg.in/yaml.v2"
@@ -47,7 +44,7 @@ type installInfo struct {
 // GetPayload builds a metadata payload every time is called.
 // Some data is collected only once, some is cached, some is collected at every call.
 func GetPayload(ctx context.Context, hostnameData hostname.Data) *Payload {
-	meta := getMeta(ctx, hostnameData)
+	meta := hostMetadataUtils.GetMeta(ctx, config.Datadog)
 	meta.Hostname = hostnameData.Hostname
 
 	p := &Payload{
@@ -80,50 +77,6 @@ func GetPayloadFromCache(ctx context.Context, hostnameData hostname.Data) *Paylo
 		return x.(*Payload)
 	}
 	return GetPayload(ctx, hostnameData)
-}
-
-// GetMeta grabs the metadata from the cache and returns it,
-// if the cache is empty, then it queries the information directly
-func GetMeta(ctx context.Context, hostnameData hostname.Data) *Meta {
-	key := buildKey("meta")
-	if x, found := cache.Cache.Get(key); found {
-		return x.(*Meta)
-	}
-	return getMeta(ctx, hostnameData)
-}
-
-// getMeta grabs the information and refreshes the cache
-func getMeta(ctx context.Context, hostnameData hostname.Data) *Meta {
-	osHostname, _ := os.Hostname()
-	tzname, _ := time.Now().Zone()
-	ec2Hostname, _ := ec2.GetHostname(ctx)
-	instanceID, _ := ec2.GetInstanceID(ctx)
-
-	var agentHostname string
-
-	if config.Datadog.GetBool("hostname_force_config_as_canonical") && hostnameData.FromConfiguration() {
-		agentHostname = hostnameData.Hostname
-	}
-
-	m := &Meta{
-		SocketHostname: osHostname,
-		Timezones:      []string{tzname},
-		SocketFqdn:     util.Fqdn(osHostname),
-		EC2Hostname:    ec2Hostname,
-		HostAliases:    cloudproviders.GetHostAliases(ctx),
-		InstanceID:     instanceID,
-		AgentHostname:  agentHostname,
-	}
-
-	if finalClusterName := kubelet.GetMetaClusterNameText(ctx, osHostname); finalClusterName != "" {
-		m.ClusterName = finalClusterName
-	}
-
-	// Cache the metadata for use in other payload
-	key := buildKey("meta")
-	cache.Cache.Set(key, m, cache.NoExpiration)
-
-	return m
 }
 
 func getNetworkMeta(ctx context.Context) *NetworkMeta {
